+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Development
 ===========
+- QuerySet limit function behaviour: Passing 0 as parameter will return all the documents in the cursor #1611
 - (Fill this out as you fix issues and develop your features).
 
 Changes in 0.14.0